requirements.txt
Flask
mysql-connector-python
paramiko
SQLAlchemy
PyMySQL

